refactor(nns): move proposal types into api crate
