+++ rs/nns/governance/api/src/lib.rs
@@ -1,1 +1,3 @@
+pub mod bitcoin;
 pub mod pb;
+pub mod subnet_rental;
